% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatainfo-package.r
\docType{package}
\name{rdatainfo}
\alias{rdatainfo}
\alias{rdatainfo-package}
\title{Tools and utilities to assist in macOS QuickLook ops}
\description{
Tools and utilities to assist in macOS QuickLook ops
}
\author{
Bob Rudis (bob@rud.is)
}
